requirements.txt
pymupdf>=1.23.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
ollama>=0.1.0
rank-bm25>=0.2.2
python-dotenv>=1.0.0
numpy>=1.24.0
